src/core/webpadize/webpadizePost.spec.ts
import { webpadPost } from '../webpadPost';
import { SteemPost } from '../SteemPost';
import { webpadizePost } from './webpadizePost';

fdescribe('#core #webpadize webpadizePost', () => {
  let steemPost: SteemPost;

  beforeEach(() => {
    steemPost = {
      id: 1,
      author: '',
      permlink: '',
      category: '',
      parent_author: '',
      parent_permlink: '',
      title: '',
      body: '',
      json_metadata: '',
      last_update: '',
      created: '',
      active: '',
      last_payout: '',
      depth: 0,
      children: 0,
      net_rshares: 0,
      abs_rshares: 0,
      vote_rshares: 0,
      children_abs_rshares: 0,
      cashout_time: '',
      max_cashout_time: '',
      total_vote_weight: 0,
      reward_weight: 0,
      total_payout_value: '',
      curator_payout_value: '',
      author_rewards: 0,
      net_votes: 0,
      root_author: '',
      root_permlink: '',
      max_accepted_payout: '1.000 SBD',
      percent_steem_dollars: 0,
      allow_replies: true,
      allow_votes: true,
      allow_curation_rewards: true,
      beneficiaries: [],
      url: '',
      root_title: '',
      pending_payout_value: '',
      total_pending_payout_value: '',
      active_votes: [],
      replies: [],
      author_reputation: '',
      promoted: '',
      body_length: 0,
      reblogged_by: []
    };
  });

  it('should return object with correct properties (webpad post)', () => {
    const webpadPost: webpadPost = webpadizePost(steemPost);

    expect(webpadPost.allowCurationRewards).toBeDefined();
    expect(webpadPost.allowVotes).toBeDefined();
    expect(webpadPost.beneficiaries).toBeDefined();
    expect(webpadPost.body).toBeDefined();
    expect(webpadPost.community).toBeDefined();
    expect(webpadPost.jsonMetadata).toBeDefined();
    expect(webpadPost.maxAcceptedPayout).toBeDefined();
    expect(webpadPost.percentSteemDollars).toBeDefined();
    expect(webpadPost.permlink).toBeDefined();
    expect(webpadPost.tags).toBeDefined();
    expect(webpadPost.thumbnailUrl).toBeDefined();
    expect(webpadPost.title).toBeDefined();
  });

  it('should NOT mutate the original steemPost object', () => {
    webpadizePost(steemPost);

    expect(steemPost).toEqual({
      id: 1,
      author: '',
      permlink: '',
      category: '',
      parent_author: '',
      parent_permlink: '',
      title: '',
      body: '',
      json_metadata: '',
      last_update: '',
      created: '',
      active: '',
      last_payout: '',
      depth: 0,
      children: 0,
      net_rshares: 0,
      abs_rshares: 0,
      vote_rshares: 0,
      children_abs_rshares: 0,
      cashout_time: '',
      max_cashout_time: '',
      total_vote_weight: 0,
      reward_weight: 0,
      total_payout_value: '',
      curator_payout_value: '',
      author_rewards: 0,
      net_votes: 0,
      root_author: '',
      root_permlink: '',
      max_accepted_payout: '1.000 SBD',
      percent_steem_dollars: 0,
      allow_replies: true,
      allow_votes: true,
      allow_curation_rewards: true,
      beneficiaries: [],
      url: '',
      root_title: '',
      pending_payout_value: '',
      total_pending_payout_value: '',
      active_votes: [],
      replies: [],
      author_reputation: '',
      promoted: '',
      body_length: 0,
      reblogged_by: []
    });
  });
});
